Add task #7

+++ src/by/bsuir/dubitskaya/task6/util/MatrixGenerator.java
@@ -1,12 +1,12 @@
 package by.bsuir.dubitskaya.task6.util;
 
 public class MatrixGenerator {
-    public static int[][] getSquareMatrix(int[] array) {
+    public static double[][] getSquareMatrix(double[] array) {
         if (array == null) {
             throw new IllegalArgumentException("Array should'n be null");
         }
 
-        int[][] matrix = new int[array.length][array.length];
+        double[][] matrix = new double[array.length][array.length];
         int index;
         for (int i = 0; i < array.length; i++) {
             for (int j = 0; j < array.length; j++) {
